+++ packages/components/src/components/icon/icon.ts
@@ -1,11 +1,10 @@
 import { css, html } from 'lit';
-import { customElement } from '../../../src/internal/register-custom-element';
+import { customElement } from '../../internal/register-custom-element';
 import { getIconLibrary, unwatchIcon, watchIcon } from './library';
 import { property, state } from 'lit/decorators.js';
 import { requestIcon } from './request';
 import { unsafeSVG } from 'lit/directives/unsafe-svg.js';
 import { watch } from '../../internal/watch';
-import componentStyles from '../../styles/component.styles';
 import SolidElement from '../../internal/solid-element';
 
 let parser: DOMParser;
@@ -132,7 +131,7 @@ export default class SdIcon extends SolidElement {
   }
 
   static styles = [
-    componentStyles,
+    ...SolidElement.styles,
     css`
       :host {
         width: 1em;