+++ packages/docs/src/stories/components/brandshape.test.stories.ts
@@ -20,7 +20,7 @@ export default {
   title: 'Components/sd-brandshape/Screenshots: sd-brandshape',
   component: 'sd-brandshape',
   tags: ['!autodocs'],
-  parameters: { ...parameters },
+  parameters: { ...parameters, controls: { disable: true } },
   args: overrideArgs([
     {
       type: 'slot',
@@ -67,7 +67,6 @@ export const Default = {
  */
 export const Shapes = {
   name: 'Shapes',
-  parameters: { controls: { exclude: ['shapes'] } },
   render: (args: any) => {
     return generateTemplate({
       axis: {
@@ -88,7 +87,6 @@ export const Shapes = {
  */
 export const Variants = {
   name: 'Variants',
-  parameters: { controls: { exclude: ['variant'] } },
   render: (args: any) => {
     return generateTemplate({
       axis: {
@@ -154,9 +152,6 @@ export const Breakpoints = {
  */
 export const Parts = {
   name: 'Parts',
-  parameters: {
-    controls: { exclude: ['base', 'content', 'shape-top', 'shape-middle', 'shape-bottom'] }
-  },
   render: (args: any) => {
     return generateTemplate({
       axis: {