+++ packages/docs/src/stories/templates/teaser-media.stories.ts
@@ -126,10 +126,6 @@ export const TeaserMediaWithCopyright = {
       .sd-copyright::after {
         padding: 0;
       }
-      .gradient-white.sd-copyright::after {
-        color: #000000;
-        text-shadow: none;
-      }
     </style>
     <div class="flex flex-col gap-12">
       <sd-teaser-media variant="gradient-dark" class="max-w-[600px]">
@@ -139,10 +135,13 @@ export const TeaserMediaWithCopyright = {
           <div slot="default" class="h-16">Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>
         </div>
       </sd-teaser-media>
-      <sd-teaser-media variant="gradient-white" class="max-w-[600px]">
+      <sd-teaser-media variant="gradient-light" class="max-w-[600px]">
         <img slot="media" src="./placeholders/images/architecture.jpg" class="aspect-video object-cover" alt="" />
         <h3 slot="headline">Headline Media Teaser</h3>
-        <div class="flex flex-col sd-copyright gradient-white" style="--copyright: '© Union Investment 2024'">
+        <div
+          class="flex flex-col sd-copyright sd-copyright--color-black gradient-white"
+          style="--copyright: '© Union Investment 2024'"
+        >
           <div slot="default" class="h-16">Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>
         </div>
       </sd-teaser-media>